pyproject.toml
[project]
name = "mlops-lab1"
version = "1.0.0"
description = "Image Classification MLOps Project with CI/CD"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.7",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "jinja2>=3.1.4",
    "httpx>=0.28.1",
    "pillow>=11.0.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pylint>=3.3.2",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["logic", "cli", "api"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pylint.messages_control]
max-line-length = 100
disable = [
    "C0103",  # Invalid name
    "C0114",  # Missing module docstring
    "C0115",  # Missing class docstring
    "C0116",  # Missing function docstring
    "R0903",  # Too few public methods
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov --cov-report=html --cov-report=term"
